% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAGenes}
\alias{TAGenes}
\title{All the 799 human tumor Associated genes with basic annotations}
\format{
A tibble data frame:
\describe{
  \item{Symbol}{HGNC Gene Symbol}
  \item{Location}{Chromosome postion}
  \item{Category}{Tumor Gene Type}
  \item{Name}{Gene biological function}
  \item{Chromosome}{Chromosome}
  }
}
\source{
http://www.binfo.ncku.edu.tw/TAG/
}
\usage{
data(TAGenes)
}
\description{
All the 799 human tumor Associated genes with basic annotations
}
\keyword{datasets}
